requirements.txt
Pillow
transformers
wheel
# flash-attn
torch
gradio